+++ packages/react-lowcode/src/ast/index.ts
@@ -1,3 +1,8 @@
 export * from './find'
 export * from './transformer'
 export * from './factory'
+
+export interface CharRange {
+    pos: number;
+    end: number;
+}